Agentic AI/research_agent/requirement.txt
langchain
langchain-groq
langchain-community
python-dotenv